extractors.[jdt,asm,spoon]; get rid of custom Launchers and downgrade Spoon
