+++ pom.xml
@@ -128,7 +128,7 @@
         <dependency>
             <groupId>fr.inria.gforge.spoon</groupId>
             <artifactId>spoon-core</artifactId>
-            <version>11.2.1-beta-3</version>
+            <version>11.2.0</version>
         </dependency>
         <dependency>
             <groupId>org.ow2.asm</groupId>
