+++ src/main/java/com/github/maracas/roseau/Roseau.java
@@ -6,9 +6,9 @@ import com.github.maracas.roseau.diff.APIDiff;
 import com.github.maracas.roseau.diff.changes.BreakingChange;
 import com.github.maracas.roseau.diff.formatter.BreakingChangesFormatter;
 import com.github.maracas.roseau.diff.formatter.BreakingChangesFormatterFactory;
-import com.github.maracas.roseau.extractors.jar.AsmAPIExtractor;
-import com.github.maracas.roseau.extractors.sources.SpoonAPIExtractor;
-import com.github.maracas.roseau.extractors.sources.SpoonUtils;
+import com.github.maracas.roseau.extractors.asm.AsmAPIExtractor;
+import com.github.maracas.roseau.extractors.spoon.SpoonAPIExtractor;
+import com.github.maracas.roseau.extractors.spoon.SpoonUtils;
 import com.google.common.base.Stopwatch;
 import org.apache.logging.log4j.Level;
 import org.apache.logging.log4j.LogManager;
