+++ src/main/java/com/github/maracas/roseau/extractors/asm/AsmClassVisitor.java
@@ -1,4 +1,4 @@
-package com.github.maracas.roseau.extractors.jar;
+package com.github.maracas.roseau.extractors.asm;
 
 import com.github.maracas.roseau.api.model.AccessModifier;
 import com.github.maracas.roseau.api.model.Annotation;
