+++ src/test/java/com/github/maracas/roseau/archunit/RoseauArchTest.java
@@ -11,11 +11,16 @@ import static com.tngtech.archunit.lang.syntax.ArchRuleDefinition.noClasses;
 class RoseauArchTest {
 	@ArchTest
 	static final ArchRule spoon_is_only_used_for_parsing =
-		noClasses().that().resideOutsideOfPackage("..roseau.extractors.sources..")
+		noClasses().that().resideOutsideOfPackage("..roseau.extractors.spoon..")
 			.should().accessClassesThat().resideInAPackage("spoon..");
 
 	@ArchTest
 	static final ArchRule asm_is_only_used_for_parsing =
-		noClasses().that().resideOutsideOfPackage("..roseau.extractors.jar..")
+		noClasses().that().resideOutsideOfPackage("..roseau.extractors.asm..")
 			.should().accessClassesThat().resideInAPackage("org.objectweb.asm..");
+
+	@ArchTest
+	static final ArchRule jdt_is_only_used_for_parsing =
+		noClasses().that().resideOutsideOfPackage("..roseau.extractors.jdt..")
+			.should().accessClassesThat().resideInAPackage("org.eclipse.jdt..");
 }
