+++ src/test/java/com/github/maracas/roseau/smoke/PopularLibrariesTestIT.java
@@ -10,10 +10,10 @@ import com.github.maracas.roseau.api.model.TypeDecl;
 import com.github.maracas.roseau.diff.APIDiff;
 import com.github.maracas.roseau.diff.changes.BreakingChange;
 import com.github.maracas.roseau.extractors.MavenClasspathBuilder;
-import com.github.maracas.roseau.extractors.jar.AsmAPIExtractor;
+import com.github.maracas.roseau.extractors.asm.AsmAPIExtractor;
 import com.github.maracas.roseau.extractors.jdt.JdtAPIExtractor;
-import com.github.maracas.roseau.extractors.sources.SpoonAPIExtractor;
-import com.github.maracas.roseau.extractors.sources.SpoonUtils;
+import com.github.maracas.roseau.extractors.spoon.SpoonAPIExtractor;
+import com.github.maracas.roseau.extractors.spoon.SpoonUtils;
 import com.google.common.base.Stopwatch;
 import com.google.common.collect.ArrayListMultimap;
 import com.google.common.collect.Multimap;
