+++ src/test/java/com/github/maracas/roseau/utils/TestUtils.java
@@ -13,9 +13,9 @@ import com.github.maracas.roseau.api.model.TypeDecl;
 import com.github.maracas.roseau.diff.APIDiff;
 import com.github.maracas.roseau.diff.changes.BreakingChange;
 import com.github.maracas.roseau.diff.changes.BreakingChangeKind;
-import com.github.maracas.roseau.extractors.jar.AsmAPIExtractor;
+import com.github.maracas.roseau.extractors.asm.AsmAPIExtractor;
 import com.github.maracas.roseau.extractors.jdt.JdtAPIExtractor;
-import com.github.maracas.roseau.extractors.sources.SpoonAPIExtractor;
+import com.github.maracas.roseau.extractors.spoon.SpoonAPIExtractor;
 import com.google.common.io.MoreFiles;
 import com.google.common.io.RecursiveDeleteOption;
 import japicmp.cmp.JApiCmpArchive;
